New: Adds the ability to get periodic updates of files being uploaded and downloaded to/from remote SFTP servers.
Please note that there is one question I need an answer to; That is, in SFTP.m writeStream:toSFTPHandle:progress: I need to know if the partially downloaded file needs to be removed at the time of cancellation or if the removal should be done by the developer invoking the method? Typically you would want to download a file to a temporary location before moving it to its final destination. In this scenario I believe it would make sense to remove the file in this logic. However, not all developers may do that. By deleting it in writeStream:toSFTPHandle:progress: there is the possibility of removing a file in its entirety… which may not be what we want.